+++ docs/docs/advanced/index.xml
@@ -6,7 +6,11 @@
 &lt;li>&lt;a href="#c-yandex--adguard-content-blocker-acb-nie-obs%C5%82uguje-wszystkich-regu%C5%82">&lt;strong>C) Yandex + AdGuard Content Blocker (ACB nie obsługuje wszystkich reguł)&lt;/strong>&lt;/a>&lt;/li>
 &lt;/ul>
 &lt;h2 id="a-yandex--nano-adblocker--nano-defender">A) Yandex + Nano Adblocker + Nano Defender&lt;/h2>
-&lt;p>(wersja wideo - już wkrótce)&lt;/p>
+&lt;hr>&lt;blockquote>&lt;details>&lt;summary>Wersja wideo&lt;/summary>&lt;br>
+&lt;div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden;">
+&lt;iframe src="https://www.youtube.com/embed/9k6tvyyowk0" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%; border:0;" allowfullscreen title="YouTube Video">&lt;/iframe>
+&lt;/div>
+&lt;/details>&lt;/blockquote>&lt;hr>
 &lt;p>Aby zainstalować Nano Adblocker dla Yandexa, powinieneś wykonać parę kroków:&lt;/p>
 &lt;ol>
 &lt;li>Po zainstalowaniu Yandex Browser możesz wejść w jego &lt;strong>ustawienia&lt;/strong> (settings) i odznaczyć &lt;strong>show Yandex Zen&lt;/strong> (Zen pokazuje newsy z różnych stron, wg. zainteresowań, historii, lokalizacji i innych czynników) i &lt;strong>show ads above Tableau&lt;/strong> (to są oczywiście reklamy), następnie wejdź w &lt;strong>notifications&lt;/strong> i odznacz &lt;strong>show Tableau in the notifications bar&lt;/strong> (Tableau to widgety najczęściej odwiedzanych stron, w tym przypadku wyświetlane na pasku). Potem wróć do głównych ustawień i ustaw &lt;strong>Shocking ad blocker&lt;/strong> na &lt;strong>Disabled&lt;/strong>.&lt;/li>
@@ -33,8 +37,16 @@
 9. Zaznacz Developer mode, dotknij Load unpacked extension, wybierz Dokumenty/MiXplorer i wejdź do katalogu, w którym masz rozpakowane rozszerzenie (Pamięć wewnętrzna/Rozszerzenia/Nano_Defender) i wybierz manifest.json.
 10. Sprawdź, czy działa. Wejdź na [antiblock.org/?p=v3&amp;demo](http://antiblock.org/?p=v3&amp;demo "http://antiblock.org/?p=v3&amp;demo"). Jeśli wyświetli się komunikat _„Please disable your ad blocker!"_, to znaczy, że być może zrobiłeś coś nie tak, jak trzeba, a jeśli nie, to gotowe!--->
 &lt;h2 id="b-yandex--ublock-origin--nano-defender">B) Yandex + uBlock Origin + Nano Defender&lt;/h2>
-&lt;p>(wersja wideo (instalowanie) ➡️ &lt;a href="https://www.youtube.com/watch?v=nmEs0J4M-7I" title="https://www.youtube.com/watch?v=nmEs0J4M-7I" target="_blank" rel="noopener">www.youtube.com/watch?v=nmEs0J4M-7I&lt;/a>)
-(wersja wideo (aktualizowanie) ➡️ &lt;a href="https://www.youtube.com/watch?v=k9t9PRDn5MU" title="https://www.youtube.com/watch?v=k9t9PRDn5MU" target="_blank" rel="noopener">www.youtube.com/watch?v=k9t9PRDn5MU&lt;/a>)&lt;/p>
+&lt;hr>&lt;blockquote>&lt;details>&lt;summary>Wersja wideo (instalowanie)&lt;/summary>&lt;br>
+&lt;div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden;">
+&lt;iframe src="https://www.youtube.com/embed/nmEs0J4M-7I" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%; border:0;" allowfullscreen title="YouTube Video">&lt;/iframe>
+&lt;/div>
+&lt;/details>&lt;/blockquote>&lt;hr>
+&lt;hr>&lt;blockquote>&lt;details>&lt;summary>Wersja wideo (aktualizowanie)&lt;/summary>&lt;br>
+&lt;div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden;">
+&lt;iframe src="https://www.youtube.com/embed/k9t9PRDn5MU" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%; border:0;" allowfullscreen title="YouTube Video">&lt;/iframe>
+&lt;/div>
+&lt;/details>&lt;/blockquote>&lt;hr>
 &lt;p>Aby zainstalować uBlock Origin (w wersji stabilnej) dla Yandexa, powinieneś wykonać parę kroków:&lt;/p>
 &lt;ol>
 &lt;li>Po zainstalowaniu Yandex Browser możesz wejść w jego &lt;strong>ustawienia&lt;/strong> (settings) i odznaczyć &lt;strong>show Yandex Zen&lt;/strong> (Zen pokazuje newsy z różnych stron, wg. zainteresowań, historii, lokalizacji i innych czynników) i &lt;strong>show ads above Tableau&lt;/strong> (to są oczywiście reklamy), następnie wejdź w &lt;strong>notifications&lt;/strong> i odznacz &lt;strong>show Tableau in the notifications bar&lt;/strong> (Tableau to widgety najczęściej odwiedzanych stron, w tym przypadku wyświetlane na pasku). Potem wróć do głównych ustawień i ustaw &lt;strong>Shocking ad blocker&lt;/strong> na &lt;strong>Disabled&lt;/strong>.&lt;/li>
@@ -151,7 +163,11 @@
 &lt;p>⚠️ &lt;strong>UWAGA!!!!&lt;/strong> ⚠️ Odmiana Pro nie zawiera niektórych funkcji, takich jak filtrowanie danych odpowiedzi.&lt;/p>
 &lt;/div>
 &lt;h2 id="iii-operaoperasubsupchrsupsubopera">&lt;strong>III. &lt;img src="https://cdnjs.cloudflare.com/ajax/libs/browser-logos/46.1.0/opera/opera_24x24.png" alt="Opera" title="Opera">&lt;sub>&lt;sup>(CHR)&lt;/sup>&lt;/sub>OPERA&lt;/strong>&lt;/h2>
-&lt;p>(wersja wideo - &lt;a href="https://www.youtube.com/watch?v=R5iB1-6B1A0" title="https://www.youtube.com/watch?v=R5iB1-6B1A0" target="_blank" rel="noopener">www.youtube.com/watch?v=R5iB1-6B1A0&lt;/a>)&lt;/p>
+&lt;hr>&lt;blockquote>&lt;details>&lt;summary>Wersja wideo&lt;/summary>&lt;br>
+&lt;div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden;">
+&lt;iframe src="https://www.youtube.com/embed/R5iB1-6B1A0" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%; border:0;" allowfullscreen title="YouTube Video">&lt;/iframe>
+&lt;/div>
+&lt;/details>&lt;/blockquote>&lt;hr>
 &lt;ol>
 &lt;li>
 &lt;p>Zainstaluj &lt;a href="https://addons.opera.com/pl/extensions/details/download-chrome-extension-9/?display=en" title="https://addons.opera.com/pl/extensions/details/download-chrome-extension-9/?display=en" target="_blank" rel="noopener">&lt;strong>Download Chrome Extension&lt;/strong>&lt;/a>.&lt;/p>
