+++ docs/docs/basic/index.xml
@@ -9,21 +9,19 @@
 &lt;/ul>
 &lt;h2 id="2-ręcznie">2. Ręcznie&lt;/h2>
 &lt;ul>
-&lt;h4 id="adblock-plus">Adblock Plus&lt;/h4>
-&lt;ul>
-&lt;li>
-&lt;p>W miejsce 3 wklejasz adres z listą filtrów np.&lt;/p>
-&lt;pre>&lt;code>https://raw.githubusercontent.com/MajkiIT/polish-ads-filter/master/polish-adblock-filters/adblock.txt
-&lt;/code>&lt;/pre>&lt;p>&lt;a href="https://raw.githubusercontent.com/MajkiIT/polish-ads-filter/master/images/563629da0110b_Beztytuu.jpg.b1a682e45ed27b33a1b85093ce1d7ef1.jpg" target="_blank" rel="noopener">&lt;img src="https://user-images.githubusercontent.com/5884000/38175733-6515e8dc-35e1-11e8-9945-06413f716eb7.jpg" alt="Adblock Plus" title="Dodawanie filtrów na przykładzie przeglądarki Maxthon w dodatku Adblock Plus">&lt;/a>&lt;/p>
-&lt;/li>
-&lt;/ul>
+&lt;!-- #### Adblock Plus
+* W miejsce 3 wklejasz adres z listą filtrów np.
+```
+https://raw.githubusercontent.com/MajkiIT/polish-ads-filter/master/polish-adblock-filters/adblock.txt
+```
+[![Adblock Plus](https://user-images.githubusercontent.com/5884000/38175733-6515e8dc-35e1-11e8-9945-06413f716eb7.jpg "Dodawanie filtrów na przykładzie przeglądarki Maxthon w dodatku Adblock Plus")](https://user-images.githubusercontent.com/5884000/38175733-6515e8dc-35e1-11e8-9945-06413f716eb7.jpg) -->
 &lt;h4 id="ublock-origin--nano-adblocker">uBlock Origin / Nano Adblocker&lt;/h4>
 &lt;ul>
 &lt;li>
-&lt;p>W miejsce 2 wklejasz adres z listą filtrów np.&lt;/p>
-&lt;pre>&lt;code>https://raw.githubusercontent.com/MajkiIT/polish-ads-filter/master/polish-adblock-filters/adblock.txt
-&lt;/code>&lt;/pre>&lt;p>&lt;img src="../../../images/wiki/uBO_add1.png" alt="ubo" title="klikamy w przycisk ustawień">]&lt;/p>
-&lt;p>&lt;img src="../../../images/wiki/uBO_add2.png" alt="ubo-settings" title="aktywujemy zakładkę zewnętrzne filtry, przewijamy stronę do końca i wklejamy link z filtrami do pola na dole, potem klikamy zastosuj">]&lt;/p>
+&lt;p>W miejsce 3 wklejasz adres z listą filtrów np.&lt;/p>
+&lt;pre>&lt;code>https://raw.githubusercontent.com/PolishFiltersTeam/KAD/master/KAD.txt
+&lt;/code>&lt;/pre>&lt;p>&lt;img src="../../../images/wiki/uBO_add1.png" alt="ubo" title="klikamy w przycisk ustawień">&lt;/p>
+&lt;p>&lt;img src="../../../images/wiki/uBO_add2.png" alt="ubo-settings" title="aktywujemy zakładkę zewnętrzne filtry, przewijamy stronę do końca i wklejamy link z filtrami do pola na dole, potem klikamy zastosuj">&lt;/p>
 &lt;/li>
 &lt;/ul>
 &lt;/ul>
@@ -32,8 +30,8 @@
 &lt;ul>
 &lt;h4 id="ublock-origin--nano-adblocker-1">uBlock Origin / Nano Adblocker&lt;/h4>
 &lt;ul>
-&lt;p>&lt;img src="../../../images/wiki/uBO_update.jpg" alt="ublock-origin-ustawienia-wtyczki">&lt;/p>
-&lt;p>&lt;img src="../../../images/wiki/uBO_update.gif" alt="aktualizacja-filtrow-w-ublock-origin">&lt;/p>
+&lt;p>&lt;img src="../../../images/wiki/uBO_update.png" alt="ublock-origin-ustawienia-rozszerzenia">&lt;/p>
+&lt;p>&lt;img src="../../../images/wiki/uBO_update2.gif" alt="aktualizacja-filtrow-w-ublocku-origin">&lt;/p>
 &lt;/ul>
 &lt;/ul></description></item><item><title>Docs: 18 sposobów na wyczyszczenie pamięci podręcznej (cache) Twojej przeglądarki internetowej (ang.)</title><link>https://majkiit.github.io/polish-ads-filter/docs/basic/18-sposob%C3%B3w-na-wyczyszczenie-pami%C4%99ci-podr%C4%99cznej-cache-twojej-przegl%C4%85darki-internetowej/</link><pubDate>Mon, 01 Jan 0001 00:00:00 +0000</pubDate><guid>https://majkiit.github.io/polish-ads-filter/docs/basic/18-sposob%C3%B3w-na-wyczyszczenie-pami%C4%99ci-podr%C4%99cznej-cache-twojej-przegl%C4%85darki-internetowej/</guid><description/></item><item><title>Docs: Jak zablokować i usunąć reklamy w Skype 7 na Windowsie?</title><link>https://majkiit.github.io/polish-ads-filter/docs/basic/skype-adblock/</link><pubDate>Mon, 01 Jan 0001 00:00:00 +0000</pubDate><guid>https://majkiit.github.io/polish-ads-filter/docs/basic/skype-adblock/</guid><description/></item><item><title>Docs: Jak zapobiec zamienianiu linków przez Google przy kliknięciu PPM?</title><link>https://majkiit.github.io/polish-ads-filter/docs/basic/search-link-fix/</link><pubDate>Wed, 26 Dec 2018 00:00:00 +0000</pubDate><guid>https://majkiit.github.io/polish-ads-filter/docs/basic/search-link-fix/</guid><description>
 &lt;p>Jak zapewne zauważyliście, jeśli wyszukacie coś przez Google i klikniecie to prawym przyciskiem myszy, aby skopiować link, to Google zamienia go np. na coś takiego:&lt;/p>
@@ -63,5 +61,5 @@
 &lt;/table>
 &lt;p>&lt;img src="http://i.imgur.com/wL13ARK.jpg" alt="Przed/po" title="Przed/po">&lt;/p></description></item><item><title>Docs: Jak stworzyć, opublikować i załadować kopię zapasową ustawień uBlocka Origin lub Nano Adblockera?</title><link>https://majkiit.github.io/polish-ads-filter/docs/basic/ubo-backup/</link><pubDate>Sun, 04 Feb 2018 00:00:00 +0000</pubDate><guid>https://majkiit.github.io/polish-ads-filter/docs/basic/ubo-backup/</guid><description>
 &lt;p>W panelu sterowania uBlocka/Nano Adblockera, w karcie ustawienia, kliknij &lt;strong>utwórz kopię zapasową&lt;/strong>. Następnie zapisz plik, a potem wrzuć go na &lt;a href="https://gist.github.com/">https://gist.github.com/&lt;/a> i podaj link.&lt;/p>
-&lt;p>&lt;img src="../../../images/wiki/Nano_backup.png" alt="2017-12-09 11_05_48-nano panel sterowania">&lt;/p>
-&lt;p>Aby załadować, pobierz plik, następnie kliknij przycisk &lt;strong>przywróć z kopii&lt;/strong>, wybierz i załaduj.&lt;/p></description></item><item><title>Docs: Jak zablokować autoodtwarzanie filmików?</title><link>https://majkiit.github.io/polish-ads-filter/docs/basic/autoplay/</link><pubDate>Mon, 01 Jan 0001 00:00:00 +0000</pubDate><guid>https://majkiit.github.io/polish-ads-filter/docs/basic/autoplay/</guid><description/></item></channel></rss>
+&lt;p>&lt;img src="../../../images/wiki/uBO_backup.png" alt="Panel sterowania uBO">&lt;/p>
+&lt;p>Aby załadować kopię, pobierz plik, następnie kliknij przycisk &lt;strong>przywróć z kopii&lt;/strong>, potem wybierz i załaduj.&lt;/p></description></item><item><title>Docs: Jak zablokować autoodtwarzanie filmików?</title><link>https://majkiit.github.io/polish-ads-filter/docs/basic/autoplay/</link><pubDate>Mon, 01 Jan 0001 00:00:00 +0000</pubDate><guid>https://majkiit.github.io/polish-ads-filter/docs/basic/autoplay/</guid><description/></item></channel></rss>
