+++ docs/sitemap.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml"><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/add-update-filter/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/yandex-adblocker/</loc><lastmod>2020-05-18T22:02:44+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/install-nano-defender/</loc><lastmod>2020-05-18T22:19:07+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/18-sposob%C3%B3w-na-wyczyszczenie-pami%C4%99ci-podr%C4%99cznej-cache-twojej-przegl%C4%85darki-internetowej/</loc></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/bezpieczny-firefox/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/skype-adblock/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/search-link-fix/</loc><lastmod>2020-05-18T22:02:44+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/dodawanie-nowej-listy-do-listy/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/ubo-backup/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/autoplay/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url></urlset>
+<?xml version="1.0" encoding="utf-8" standalone="yes"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml"><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/add-update-filter/</loc><lastmod>2020-05-19T15:06:58+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/yandex-adblocker/</loc><lastmod>2020-05-19T15:06:58+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/install-nano-defender/</loc><lastmod>2020-05-19T15:06:58+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/18-sposob%C3%B3w-na-wyczyszczenie-pami%C4%99ci-podr%C4%99cznej-cache-twojej-przegl%C4%85darki-internetowej/</loc></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/bezpieczny-firefox/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/skype-adblock/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/search-link-fix/</loc><lastmod>2020-05-18T22:02:44+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/advanced/dodawanie-nowej-listy-do-listy/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/ubo-backup/</loc><lastmod>2020-05-19T15:06:58+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/basic/autoplay/</loc><lastmod>2020-03-29T22:42:36+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/docs/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url><url><loc>https://majkiit.github.io/polish-ads-filter/</loc><lastmod>2020-03-30T14:49:00+02:00</lastmod></url></urlset>
